Added validations to admin

+++ TollSystem/TollSystem/Core/UseCases/TollStation/Commands/DeleteTollStationCommand.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using System.Windows;
 using TollSystem.Core.Entities;
 using TollSystem.Core.Services;
 using TollSystem.DesktopHost.Controllers;
@@ -18,6 +19,11 @@ namespace TollSystem.Commands
 
         public override void Execute(object parameter)
         {
+            if (_viewModel.SelectedStation is null)
+            {
+                MessageBox.Show("Izaberite stanicu !");
+                return;
+            }
             TollStationEntity station = _viewModel.SelectedStation.Station;
             IDeleteTollStationService deleteStation = new DeleteTollStationService(ServiceContainer.TollStationRepositoryService, ServiceContainer.TollStationModelService);
             deleteStation.DeleteTollStation(station);
